refactor: separate LoginForm component from the login page
