+++ frontend/src/app/components/features/loginForm/LoginForm.module.scss
@@ -1,0 +1,14 @@
+.buttonWrapper {
+  display: flex;
+  justify-content: center;
+  margin: 1rem;
+}
+
+.icon {
+  width: 1.2rem;
+  height: 1.2rem;
+}
+
+.form {
+  width: 100%;
+}
