+++ frontend/src/app/customers/login/page.module.scss
@@ -46,17 +46,4 @@
   left: 1rem;
 }
 
-.buttonWrapper {
-  display: flex;
-  justify-content: center;
-  margin: 1rem;
-}
 
-.icon {
-  width: 1.2rem;
-  height: 1.2rem;
-}
-
-.form {
-  width: 100%;
-}
